Only change default network settings on Rackspace Cloud
Having this package installed makes it very frustrating to use the nova
cli with other cloud providers, because some default network settings
are set that don't make sense anywhere but the Rackspace Cloud. This
patch adds a simple check to see if the client is actually being used
against the Rackspace Cloud.